Refactor automated workflow into modular package
