+++ automated_refactoring_workflow.py
@@ -1,514 +1,8 @@
 #!/usr/bin/env python3
-"""
-Automated Refactoring Workflow Implementation - REFACTOR-002
-==========================================================
+"""CLI entry point for the automated refactoring workflow system."""
 
-This module implements automated workflows for refactoring operations,
-including workflow automation, validation system, and reliability testing.
+from tools.refactoring_workflow.execution import main
 
-Contract: REFACTOR-002: Automated Refactoring Workflow Implementation - 200 points
-Agent: Agent-5 (SPRINT ACCELERATION REFACTORING TOOL PREPARATION MANAGER)
-Status: IN PROGRESS - Workflow Design and Implementation
-
-Requirements:
-1. Design automated workflows
-2. Implement workflow automation
-3. Create validation system
-4. Test workflow reliability
-
-Deliverables:
-1. Automated refactoring workflows
-2. Validation system
-3. Workflow reliability results
-"""
-
-import json
-import logging
-import time
-import threading
-from datetime import datetime, timezone
-from pathlib import Path
-from typing import Any, Dict, List, Optional, Callable
-from dataclasses import dataclass, field
-from enum import Enum
-import queue
-import concurrent.futures
-
-# ============================================================================
-# WORKFLOW TYPES AND STATES
-# ============================================================================
-
-class WorkflowType(Enum):
-    """Types of automated refactoring workflows"""
-    CODE_ANALYSIS = "code_analysis"
-    MODULARIZATION = "modularization"
-    DUPLICATION_ELIMINATION = "duplication_elimination"
-    PERFORMANCE_OPTIMIZATION = "performance_optimization"
-    CODE_STANDARDS_ENFORCEMENT = "code_standards_enforcement"
-    ARCHITECTURE_REFACTORING = "architecture_refactoring"
-
-class WorkflowState(Enum):
-    """Workflow execution states"""
-    PENDING = "pending"
-    RUNNING = "running"
-    VALIDATING = "validating"
-    COMPLETED = "completed"
-    FAILED = "failed"
-    ROLLED_BACK = "rolled_back"
-
-class ValidationLevel(Enum):
-    """Validation levels for refactoring operations"""
-    BASIC = "basic"
-    STANDARD = "standard"
-    COMPREHENSIVE = "comprehensive"
-    EXHAUSTIVE = "exhaustive"
-
-# ============================================================================
-# WORKFLOW COMPONENTS
-# ============================================================================
-
-@dataclass
-class WorkflowStep:
-    """Individual step in a refactoring workflow"""
-    id: str
-    name: str
-    description: str
-    action: Callable
-    validation_rules: List[Dict[str, Any]]
-    dependencies: List[str] = field(default_factory=list)
-    timeout: int = 300  # seconds
-    retry_count: int = 3
-
-@dataclass
-class WorkflowExecution:
-    """Workflow execution instance"""
-    id: str
-    workflow_type: WorkflowType
-    target_files: List[str]
-    parameters: Dict[str, Any]
-    state: WorkflowState
-    start_time: str
-    end_time: Optional[str] = None
-    steps_completed: List[str] = field(default_factory=list)
-    validation_results: Dict[str, Any] = field(default_factory=dict)
-    error_log: List[str] = field(default_factory=list)
-    performance_metrics: Dict[str, Any] = field(default_factory=dict)
-
-# ============================================================================
-# AUTOMATED WORKFLOW ENGINE
-# ============================================================================
-
-class AutomatedWorkflowEngine:
-    """Main engine for automated refactoring workflows"""
-    
-    def __init__(self):
-        self.workflows = {}
-        self.execution_queue = queue.Queue()
-        self.active_executions = {}
-        self.workflow_history = []
-        self.validation_system = WorkflowValidationSystem()
-        self.performance_monitor = WorkflowPerformanceMonitor()
-        
-        # Start workflow execution thread
-        self.execution_thread = threading.Thread(target=self._execution_worker, daemon=True)
-        self.execution_thread.start()
-        
-        self._initialize_workflows()
-    
-    def _initialize_workflows(self):
-        """Initialize predefined refactoring workflows"""
-        # Code Analysis Workflow
-        self.workflows[WorkflowType.CODE_ANALYSIS] = {
-            "name": "Automated Code Analysis",
-            "description": "Comprehensive code analysis and optimization workflow",
-            "steps": [
-                WorkflowStep(
-                    id="analyze_code_structure",
-                    name="Code Structure Analysis",
-                    description="Analyze code structure and identify optimization opportunities",
-                    action=self._analyze_code_structure,
-                    validation_rules=[{"type": "structure_valid", "required": True}]
-                ),
-                WorkflowStep(
-                    id="identify_patterns",
-                    name="Pattern Identification",
-                    description="Identify common patterns and anti-patterns",
-                    action=self._identify_patterns,
-                    validation_rules=[{"type": "patterns_found", "min_count": 1}]
-                ),
-                WorkflowStep(
-                    id="generate_recommendations",
-                    name="Optimization Recommendations",
-                    description="Generate optimization recommendations",
-                    action=self._generate_recommendations,
-                    validation_rules=[{"type": "recommendations_generated", "min_count": 1}]
-                )
-            ]
-        }
-        
-        # Modularization Workflow
-        self.workflows[WorkflowType.MODULARIZATION] = {
-            "name": "Automated Modularization",
-            "description": "Break down large files into smaller, maintainable modules",
-            "steps": [
-                WorkflowStep(
-                    id="analyze_file_complexity",
-                    name="File Complexity Analysis",
-                    description="Analyze file complexity and identify modularization opportunities",
-                    action=self._analyze_file_complexity,
-                    validation_rules=[{"type": "complexity_analyzed", "required": True}]
-                ),
-                WorkflowStep(
-                    id="design_modular_structure",
-                    name="Modular Structure Design",
-                    description="Design modular structure for complex files",
-                    action=self._design_modular_structure,
-                    validation_rules=[{"type": "structure_designed", "required": True}]
-                ),
-                WorkflowStep(
-                    id="implement_modularization",
-                    name="Modularization Implementation",
-                    description="Implement the modularization plan",
-                    action=self._implement_modularization,
-                    validation_rules=[{"type": "modularization_implemented", "required": True}]
-                )
-            ]
-        }
-    
-    def execute_workflow(self, workflow_type: WorkflowType, target_files: List[str], 
-                        parameters: Dict[str, Any] = None) -> str:
-        """Execute a refactoring workflow"""
-        if workflow_type not in self.workflows:
-            raise ValueError(f"Unknown workflow type: {workflow_type}")
-        
-        execution_id = f"{workflow_type.value}_{int(time.time())}"
-        execution = WorkflowExecution(
-            id=execution_id,
-            workflow_type=workflow_type,
-            target_files=target_files,
-            parameters=parameters or {},
-            state=WorkflowState.PENDING,
-            start_time=datetime.now(timezone.utc).isoformat()
-        )
-        
-        # Add to execution queue
-        self.execution_queue.put(execution)
-        self.active_executions[execution_id] = execution
-        
-        logging.info(f"Workflow {workflow_type.value} queued for execution: {execution_id}")
-        return execution_id
-    
-    def _execution_worker(self):
-        """Background worker for executing workflows"""
-        while True:
-            try:
-                execution = self.execution_queue.get(timeout=1)
-                self._execute_workflow(execution)
-            except queue.Empty:
-                continue
-            except Exception as e:
-                logging.error(f"Workflow execution error: {e}")
-    
-    def _execute_workflow(self, execution: WorkflowExecution):
-        """Execute a single workflow"""
-        try:
-            execution.state = WorkflowState.RUNNING
-            workflow = self.workflows[execution.workflow_type]
-            
-            logging.info(f"Starting workflow execution: {execution.id}")
-            
-            # Execute each step
-            for step in workflow["steps"]:
-                if not self._can_execute_step(step, execution):
-                    continue
-                
-                try:
-                    # Execute step
-                    step_result = step.action(execution)
-                    
-                    # Validate step result
-                    validation_result = self.validation_system.validate_step(step, step_result)
-                    
-                    if validation_result["valid"]:
-                        execution.steps_completed.append(step.id)
-                        logging.info(f"Step {step.name} completed successfully")
-                    else:
-                        raise Exception(f"Step validation failed: {validation_result['errors']}")
-                        
-                except Exception as e:
-                    logging.error(f"Step {step.name} failed: {e}")
-                    execution.error_log.append(f"Step {step.name}: {str(e)}")
-                    execution.state = WorkflowState.FAILED
-                    return
-            
-            # Final validation
-            execution.state = WorkflowState.VALIDATING
-            final_validation = self.validation_system.validate_workflow(execution)
-            
-            if final_validation["valid"]:
-                execution.state = WorkflowState.COMPLETED
-                execution.end_time = datetime.now(timezone.utc).isoformat()
-                logging.info(f"Workflow {execution.id} completed successfully")
-            else:
-                execution.state = WorkflowState.FAILED
-                execution.error_log.append(f"Final validation failed: {final_validation['errors']}")
-            
-        except Exception as e:
-            execution.state = WorkflowState.FAILED
-            execution.error_log.append(f"Workflow execution failed: {str(e)}")
-            logging.error(f"Workflow execution failed: {e}")
-        
-        finally:
-            # Record performance metrics
-            execution.performance_metrics = self.performance_monitor.get_execution_metrics(execution)
-            
-            # Add to history
-            self.workflow_history.append(execution)
-            
-            # Remove from active executions
-            if execution.id in self.active_executions:
-                del self.active_executions[execution.id]
-    
-    def _can_execute_step(self, step: WorkflowStep, execution: WorkflowExecution) -> bool:
-        """Check if a step can be executed"""
-        # Check dependencies
-        for dependency in step.dependencies:
-            if dependency not in execution.steps_completed:
-                return False
-        return True
-    
-    # Workflow step implementations
-    def _analyze_code_structure(self, execution: WorkflowExecution) -> Dict[str, Any]:
-        """Analyze code structure"""
-        # Implementation would analyze target files
-        return {"structure_analyzed": True, "complexity_score": 0.75}
-    
-    def _identify_patterns(self, execution: WorkflowExecution) -> Dict[str, Any]:
-        """Identify code patterns"""
-        return {"patterns_found": 5, "anti_patterns": 2}
-    
-    def _generate_recommendations(self, execution: WorkflowExecution) -> Dict[str, Any]:
-        """Generate optimization recommendations"""
-        return {"recommendations_generated": 8, "priority": "high"}
-    
-    def _analyze_file_complexity(self, execution: WorkflowExecution) -> Dict[str, Any]:
-        """Analyze file complexity"""
-        return {"complexity_analyzed": True, "files_analyzed": len(execution.target_files)}
-    
-    def _design_modular_structure(self, execution: WorkflowExecution) -> Dict[str, Any]:
-        """Design modular structure"""
-        return {"structure_designed": True, "modules_planned": 3}
-    
-    def _implement_modularization(self, execution: WorkflowExecution) -> Dict[str, Any]:
-        """Implement modularization"""
-        return {"modularization_implemented": True, "files_created": 3}
-
-# ============================================================================
-# WORKFLOW VALIDATION SYSTEM
-# ============================================================================
-
-class WorkflowValidationSystem:
-    """System for validating workflow steps and results"""
-    
-    def __init__(self):
-        self.validation_rules = {}
-        self._initialize_validation_rules()
-    
-    def _initialize_validation_rules(self):
-        """Initialize validation rules for different step types"""
-        self.validation_rules = {
-            "structure_valid": lambda result: result.get("structure_analyzed", False),
-            "patterns_found": lambda result: result.get("patterns_found", 0) >= 1,
-            "recommendations_generated": lambda result: result.get("recommendations_generated", 0) >= 1,
-            "complexity_analyzed": lambda result: result.get("complexity_analyzed", False),
-            "structure_designed": lambda result: result.get("structure_designed", False),
-            "modularization_implemented": lambda result: result.get("modularization_implemented", False)
-        }
-    
-    def validate_step(self, step: WorkflowStep, result: Dict[str, Any]) -> Dict[str, Any]:
-        """Validate a workflow step result"""
-        validation_result = {
-            "valid": True,
-            "errors": [],
-            "warnings": [],
-            "timestamp": datetime.now(timezone.utc).isoformat()
-        }
-        
-        for rule in step.validation_rules:
-            rule_type = rule["type"]
-            if rule_type in self.validation_rules:
-                try:
-                    if not self.validation_rules[rule_type](result):
-                        validation_result["valid"] = False
-                        validation_result["errors"].append(f"Validation rule failed: {rule_type}")
-                except Exception as e:
-                    validation_result["valid"] = False
-                    validation_result["errors"].append(f"Validation error for {rule_type}: {str(e)}")
-            else:
-                validation_result["warnings"].append(f"Unknown validation rule: {rule_type}")
-        
-        return validation_result
-    
-    def validate_workflow(self, execution: WorkflowExecution) -> Dict[str, Any]:
-        """Validate complete workflow execution"""
-        validation_result = {
-            "valid": True,
-            "errors": [],
-            "warnings": [],
-            "timestamp": datetime.now(timezone.utc).isoformat()
-        }
-        
-        # Check if all steps completed
-        workflow = execution.workflow_type
-        expected_steps = len(execution.workflow_type.steps) if hasattr(execution.workflow_type, 'steps') else 0
-        
-        if len(execution.steps_completed) < expected_steps:
-            validation_result["valid"] = False
-            validation_result["errors"].append(f"Not all steps completed: {len(execution.steps_completed)}/{expected_steps}")
-        
-        # Check for errors
-        if execution.error_log:
-            validation_result["valid"] = False
-            validation_result["errors"].extend(execution.error_log)
-        
-        return validation_result
-
-# ============================================================================
-# WORKFLOW PERFORMANCE MONITOR
-# ============================================================================
-
-class WorkflowPerformanceMonitor:
-    """Monitor workflow performance and reliability"""
-    
-    def __init__(self):
-        self.performance_metrics = {}
-    
-    def get_execution_metrics(self, execution: WorkflowExecution) -> Dict[str, Any]:
-        """Get performance metrics for a workflow execution"""
-        if execution.end_time and execution.start_time:
-            start_time = datetime.fromisoformat(execution.start_time.replace('Z', '+00:00'))
-            end_time = datetime.fromisoformat(execution.end_time.replace('Z', '+00:00'))
-            duration = (end_time - start_time).total_seconds()
-        else:
-            duration = 0
-        
-        return {
-            "duration_seconds": duration,
-            "steps_completed": len(execution.steps_completed),
-            "success_rate": 1.0 if execution.state == WorkflowState.COMPLETED else 0.0,
-            "error_count": len(execution.error_log),
-            "timestamp": datetime.now(timezone.utc).isoformat()
-        }
-
-# ============================================================================
-# MAIN WORKFLOW SYSTEM
-# ============================================================================
-
-class AutomatedRefactoringWorkflowSystem:
-    """Main system for automated refactoring workflows"""
-    
-    def __init__(self):
-        self.workflow_engine = AutomatedWorkflowEngine()
-        self.validation_system = WorkflowValidationSystem()
-        self.performance_monitor = WorkflowPerformanceMonitor()
-    
-    def execute_code_analysis_workflow(self, target_files: List[str], 
-                                     parameters: Dict[str, Any] = None) -> str:
-        """Execute code analysis workflow"""
-        return self.workflow_engine.execute_workflow(
-            WorkflowType.CODE_ANALYSIS, target_files, parameters
-        )
-    
-    def execute_modularization_workflow(self, target_files: List[str], 
-                                      parameters: Dict[str, Any] = None) -> str:
-        """Execute modularization workflow"""
-        return self.workflow_engine.execute_workflow(
-            WorkflowType.MODULARIZATION, target_files, parameters
-        )
-    
-    def get_workflow_status(self, execution_id: str) -> Optional[WorkflowExecution]:
-        """Get status of a workflow execution"""
-        return self.workflow_engine.active_executions.get(execution_id)
-    
-    def get_workflow_history(self) -> List[WorkflowExecution]:
-        """Get workflow execution history"""
-        return self.workflow_engine.workflow_history
-    
-    def get_performance_report(self) -> Dict[str, Any]:
-        """Get comprehensive performance report"""
-        history = self.workflow_engine.workflow_history
-        if not history:
-            return {"message": "No workflow executions yet"}
-        
-        total_executions = len(history)
-        successful_executions = len([h for h in history if h.state == WorkflowState.COMPLETED])
-        failed_executions = len([h for h in history if h.state == WorkflowState.FAILED])
-        
-        total_duration = sum(h.performance_metrics.get("duration_seconds", 0) for h in history)
-        avg_duration = total_duration / total_executions if total_executions > 0 else 0
-        
-        return {
-            "total_executions": total_executions,
-            "successful_executions": successful_executions,
-            "failed_executions": failed_executions,
-            "success_rate": successful_executions / total_executions if total_executions > 0 else 0,
-            "average_duration_seconds": avg_duration,
-            "total_duration_seconds": total_duration,
-            "timestamp": datetime.now(timezone.utc).isoformat()
-        }
-
-# ============================================================================
-# CLI INTERFACE
-# ============================================================================
-
-def main():
-    """Main CLI interface for the Automated Refactoring Workflow System"""
-    print("🚀 AUTOMATED REFACTORING WORKFLOW SYSTEM")
-    print("=" * 50)
-    print("Contract: REFACTOR-002 - Automated Refactoring Workflow Implementation")
-    print("Agent: Agent-5 - SPRINT ACCELERATION REFACTORING TOOL PREPARATION MANAGER")
-    print("=" * 50)
-    
-    # Initialize system
-    workflow_system = AutomatedRefactoringWorkflowSystem()
-    
-    # Example usage
-    print("\n📋 Available Workflows:")
-    print("1. Code Analysis Workflow")
-    print("2. Modularization Workflow")
-    
-    # Execute example workflow
-    print("\n🔄 Executing Code Analysis Workflow...")
-    execution_id = workflow_system.execute_code_analysis_workflow(
-        target_files=["src/core/unified_data_source_consolidation.py"],
-        parameters={"analysis_level": "comprehensive"}
-    )
-    
-    print(f"✅ Workflow started: {execution_id}")
-    
-    # Wait for completion
-    print("⏳ Waiting for workflow completion...")
-    while True:
-        status = workflow_system.get_workflow_status(execution_id)
-        if status and status.state in [WorkflowState.COMPLETED, WorkflowState.FAILED]:
-            break
-        time.sleep(1)
-    
-    # Show results
-    print(f"\n📊 Workflow Status: {status.state.value}")
-    if status.state == WorkflowState.COMPLETED:
-        print("✅ Workflow completed successfully!")
-        print(f"Steps completed: {len(status.steps_completed)}")
-    else:
-        print("❌ Workflow failed!")
-        print(f"Errors: {status.error_log}")
-    
-    # Performance report
-    print("\n📈 Performance Report:")
-    performance = workflow_system.get_performance_report()
-    for key, value in performance.items():
-        if key != "timestamp":
-            print(f"  {key}: {value}")
 
 if __name__ == "__main__":
     main()
